move function to fix circular import
