+++ src/oge/helpers.py
@@ -6,7 +6,7 @@ from geopy.exc import GeocoderUnavailable
 from timezonefinder import TimezoneFinder
 from urllib3.exceptions import ReadTimeoutError
 
-from oge.column_checks import get_dtypes, apply_dtypes
+from oge.column_checks import get_dtypes, apply_dtypes, DATA_COLUMNS
 from oge.constants import (
     earliest_data_year,
     latest_validated_year,
@@ -307,6 +307,116 @@ def assign_fleet_to_subplant_data(
         )
 
     return subplant_data
+
+
+def combine_subplant_data(
+    cems: pd.DataFrame,
+    partial_cems_subplant: pd.DataFrame,
+    partial_cems_plant: pd.DataFrame,
+    eia_data: pd.DataFrame,
+    resolution: str,
+    validate: bool = True,
+) -> pd.DataFrame:
+    """Combines subplant-level data from multiple sources into a single dataframe.
+    Data can be returned at either the monthly or hourly resolution.
+
+    When passing hourly data in later in the pipeline, we only want to combine cems data
+    so we can pass an empty dataframe for eia_data.
+
+    Args:
+        cems (pd.DataFrame): _description_
+        partial_cems_subplant (pd.DataFrame): _description_
+        partial_cems_plant (pd.DataFrame): _description_
+        eia_data (pd.DataFrame): _description_
+        resolution (str): _description_
+        validate (bool, optional): _description_. Defaults to True.
+
+    Raises:
+        UserWarning: _description_
+
+    Returns:
+        pd.DataFrame: _description_
+    """
+
+    KEY_COLUMNS = [
+        "plant_id_eia",
+        "subplant_id",
+        "report_date",
+    ]
+    if resolution == "hourly":
+        KEY_COLUMNS += ["datetime_utc"]
+    elif resolution == "monthly":
+        pass
+    else:
+        raise UserWarning(
+            f"`resolution` must be 'monthly' or 'hourly'. '{resolution}' specified"
+        )
+
+    ALL_COLUMNS = KEY_COLUMNS + DATA_COLUMNS
+
+    if validate:
+        validation.ensure_non_overlapping_data_from_all_sources(
+            cems, partial_cems_subplant, partial_cems_plant, eia_data
+        )
+
+    # group data by plant-hour and filter columns
+    cems = (
+        cems.groupby(
+            KEY_COLUMNS,
+            dropna=False,
+        )
+        .sum(numeric_only=True)
+        .reset_index()[[col for col in cems.columns if col in ALL_COLUMNS]]
+    )
+    # don't group if there is no data in the dataframe
+    if len(partial_cems_subplant) > 0:
+        partial_cems_subplant = (
+            partial_cems_subplant.groupby(
+                KEY_COLUMNS,
+                dropna=False,
+            )
+            .sum(numeric_only=True)
+            .reset_index()[
+                [col for col in partial_cems_subplant.columns if col in ALL_COLUMNS]
+            ]
+        )
+    if len(partial_cems_plant) > 0:
+        partial_cems_plant = (
+            partial_cems_plant.groupby(
+                KEY_COLUMNS,
+                dropna=False,
+            )
+            .sum(numeric_only=True)
+            .reset_index()[
+                [col for col in partial_cems_plant.columns if col in ALL_COLUMNS]
+            ]
+        )
+    eia_data = (
+        eia_data.groupby(
+            KEY_COLUMNS,
+            dropna=False,
+        )
+        .sum(numeric_only=True)
+        .reset_index()[[col for col in eia_data.columns if col in ALL_COLUMNS]]
+    )
+
+    # concat together
+    combined_subplant_data = pd.concat(
+        [cems, partial_cems_subplant, partial_cems_plant, eia_data],
+        axis=0,
+        ignore_index=True,
+        copy=False,
+    )
+
+    # groupby subplant after combining in case subplant reported multiple places
+    combined_subplant_data = (
+        combined_subplant_data.groupby(KEY_COLUMNS, dropna=False).sum().reset_index()
+    )
+
+    # re-order the columns
+    combined_subplant_data = combined_subplant_data[ALL_COLUMNS]
+
+    return combined_subplant_data
 
 
 def create_plant_ba_table(year: int) -> pd.DataFrame:
